refactor: Optimize for Qt6 modules[100254]

+++ qmmp/src/plugins/Light/lightwaveform/lightwaveform.cpp
@@ -617,7 +617,7 @@ void LightWaveForm::mouseMoveEvent(QMouseEvent *e)
 {
     if(m_seekPos >= 0)
     {
-        m_seekPos = qBound(0, QtMouseX(e), width());
+        m_seekPos = qBound(0, static_cast<int>(QtMouseX(e)), width());
         update();
     }
 }
